Update patch set 1
Patch Set 1:

biggest open question for me is how this plugin would be configured. Since it takes info from one (repo,branch,file) and sticks it into another (repo,branch), it seems logical to have this be global, OTOH, that would require restarting the server to change the config (which may be annoying). 

Thoughts?

Patch-set: 1