Print the canonicalWebUrl for warnings.
SuperManifest is deployed on googlesource.com, and the log messages
are the only way to get diagnostics. In general, log messages on
googlesource.com include the hostname of the problematic site by
modifying the thread names. Unfortunately, updates in SuperManifest
are driven by ReceiveCommits, which runs in the executor configured by
AsyncReceiveCommits.

Tasks in this executor have no non-opensource call frames, and
therefor we have no injection points to set the hostname.

Change-Id: I41795e50986d7164e2362f98e4bef37eb1de395a