Allow specifying which groups to subscribe to
JGit's RepoCommand defaults to tracking groups all,-notdefault.  Most
users of this plugin are likely to want to track all groups instead,
so provide a knob to allow requesting that.

If the value `all` proves overwhelmingly popular in practice, then it
can be made the default in JGit and this setting could be removed.

Change-Id: Ibddb5fdbaa9d57e6bce12affd39e14c73ee481e5
Reported-by: Jeff Gaston <jeffrygaston@google.com>