Fix a bug that causing JiriUpdater treating remote project as local
Before this patch, JiriUpdater determines whether a repo is within
local gerrit by opening a local repo with the same name. This
behavior is causing issues if a remote repo has the same name as a
local repo. This patch fixed this issue by matching the hostnames
of gerrit server and repo url.

Change-Id: I0c6ff995a6b7f1b3af0b55ec872cbc87f1059bb3