repo.ref expects ref but gets sha1
Fixed, replaced repo.ref with repo.resolve which in fact expects sha1
and gets sha1.

The function readFileWithMode which was supposed to be called only with
ref, can now be called with sha1 as well. Will work either way since
findRef will return null when called with sha1, and then we can use
repo.resolve to find the objectId using sha1. When called with ref,
findRef works.

Change-Id: Iaee9db3bbe33f1a1340d94a68ffdd93bdd6489c7