Normalize repo opening paths
openRepository() doesn't know if it is receiving a URL or a repoName
(the host-relative name of the project). It just tries to clean up the
canonicalWebUrl and assumes it is a repoName. This is wrong for e.g.
external URLs.

Besides, the URL->repoName translation is inconsistent: sha1() just
takes the path of the incoming URL (and calls openRepository()) while
readFileWithMode passes through the URL. An external URL can open the
wrong repo in the first case and nothing on the second.

Make the caller explicit if it is sending a URI or a repoName. For plain
openRepository() assume it is repoName. It seems all callers use
repoName (otherwise we need to add some guessing code).

Translate URIs to repoNames consistently (even if wrongly).

This paves the way to plug a cross-host supporting implementation.

Change-Id: Ia2a925764bb9c3305a434240d44873b31c6e0c75