Simplify the URI/repoName handling
The methods in the GerritRemoteReader can receive URIs or repoNames
(a host-relative name of the project). The handling of the URIs is
inconsistent: sha1() takes the path before calling openRepository()
while readFileWithMode() passes through the whole URL. This is wrong
e.g. for external URLs. An outter host URL can open the wrong repo in
sha1() and nothing in readFileWithMode().

Make internal methods to open by URI or name. The public methods
must decide what to invoke. This makes clear what is coming and gives a
unique place for the URI -> repoName translation.

This also paves the way to plug a repository opener that supports
cross-host URLs.

Change-Id: Ia2a925764bb9c3305a434240d44873b31c6e0c75