Inject the SuperManifestRepoManager
When a manifest contains a project in a second host, the manifest parser
generates a correct .gitmodules but when reading the SHA1, the
SuperManifestRepoManager returns <this-host>/project. In some cases this
produces an inconsistent superproject: the .gitmodules points to a repo
but the SHA1 in the gitlink doesn't exist there.

Inject the implementation of the SuperManifestRepoManager. Providers can
offer an implementation that reads correctly cross-host if their backend
supports it.

Change-Id: I88187f883b59b7b9e7428963865b790ec8e87ca3