SuperManifestRepoManager: Take the path if the URL is not canonical
In some cases the canonical URL doesn't match with the repo URL and
relying on the path returns the right result (at the risk of opening the
wrong repo if the host is different).

This was the behaviour before and the change broke some existing
manifest projects.

Revert to this (wrong) behavior until we can have a proper cross-host
implementation.

Change-Id: Id1f95c88d2f3383fbb0801c65253fa1dcb0b878d