SuperManifestRefUpdatedListener: inline the update method
The update() method has a boolean parameter to let the exceptions pass
through or not. This makes the exception handling unnecessarily
difficult to read. The method is only called from two places (with the
param set as true and false on each).

Inline the #update() method to the callers. This moves the exception
handling to the same level for both paths (REST call/plugin call) and
makes clearer what happens on exceptions: the REST call let the
exception go and stops, the plugin still tries with remaning confs.

Change-Id: I17933da33437dea4cbbe179b1ef4b6ffeb0655cf