RepoUpdate: store in destination the source manifest revision
Right now it is not possible to know the specific revision of a manifest
that defined the destination repository. This is problematic for
debugging and to detect mismatches e.g. when the manifest is updated but
the translation to superproject failed.

Save in the destination repository what manifest defined it. We store
"$repo $ref $commit" of the source manifest in an arbitrary file
(".supermanifest-src").

This uses the RepoCommand#addToDestination method introduced in JGit
6.1.

Change-Id: Iad66199d40828143f4764421247714b6f0625233