SuperManifestRefUpdate: verify no overlapping writes in same update
When confs support flexible refspects, we can find that two confs
match a ref and both write to the same destination. This is
undesirable for predictability and performance.

After finding the relevant configurations for a manifest update,
verify that not two of them write to the same destination. If there is
an overlap, the error out and skip the writing (admins should fix
the conf).

Change-Id: Ia0ed7b8eb7f9506909ef14b2e71141575d4b813f