RepoUpdater: allow to set recordRemoteBranch via config
The RepoCommand has a recordRemoteBranch option, but it is not read
from the configuration (it is always set to true). The "branch"
attribute controls also the autoupdate of submodules and in some
branches we want to disable that.

Read the value from the conf and pass it to the RepoUpdater. Default
to true to keep the current behavior.

Change-Id: I728853f69bf3c21c882b3eae17e4eeb7b2cc3621